@@ -1,10 +1,10 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-# Packages may add whatever they like to this file, but
-# should keep this content at the top.
-# ----------------------------------------------------------------------------
-from ._astropy_init import *   # noqa
-# ----------------------------------------------------------------------------
-
 from specreduce.core import *  # noqa
 from specreduce.wavelength_calibration import * # noqa
+
+
+try:
+    from .version import version as __version__
+except ImportError:
+    __version__ = ''
