@@ -17,7 +17,7 @@ def test_pypeit_single():
         'Source',
         'amplitude',
         'ion',
-        'wave'
+        'wavelength'
     ]
 
 
@@ -27,6 +27,17 @@ def test_pypeit_list():
     Test to load and combine a set of linelists from ``pypeit`` by passing a list.
     """
     line_tab = load_pypeit_calibration_lines(['HeI', 'NeI'], cache=True, show_progress=False)
+    assert line_tab is not None
+    assert "HeI" in line_tab['ion']
+    assert "NeI" in line_tab['ion']
+
+
+@pytest.mark.remote_data
+def test_pypeit_comma_list():
+    """
+    Test to load and combine a set of linelists from ``pypeit`` by passing a comma-separated list.
+    """
+    line_tab = load_pypeit_calibration_lines("HeI, NeI", cache=True, show_progress=False)
     assert line_tab is not None
     assert "HeI" in line_tab['ion']
     assert "NeI" in line_tab['ion']
@@ -48,7 +59,9 @@ def test_pypeit_input_validation():
     Check that bad inputs for ``pypeit`` linelists raise the appropriate warnings and exceptions
     """
     with pytest.raises(ValueError, match='.*Invalid calibration lamps specification.*'):
-        load_pypeit_calibration_lines({}, cache=True, show_progress=False)
+        _ = load_pypeit_calibration_lines({}, cache=True, show_progress=False)
 
-    with pytest.warns(UserWarning, match="ArIII not in the list of supported calibration line lists"):  # noqa: E501
-        load_pypeit_calibration_lines(['HeI', 'ArIII'], cache=True, show_progress=False)
+    with pytest.warns() as record:
+        _ = load_pypeit_calibration_lines(['HeI', 'ArIII'], cache=True, show_progress=False)
+        if not record:
+            pytest.fails("Expected warning about nonexistant linelist for ArIII.")
